chore: split up ban command into multiple files.
