+++ interactions/ban/ban.go
@@ -10,7 +10,6 @@ import (
 	"github.com/disgoorg/json"
 
 	"github.com/NLLCommunity/heimdallr/interactions"
-	"github.com/NLLCommunity/heimdallr/model"
 	"github.com/NLLCommunity/heimdallr/utils"
 )
 
@@ -31,128 +30,9 @@ var BanCommand = discord.SlashCommandCreate{
 	DMPermission:             utils.Ref(false),
 	DefaultMemberPermissions: json.NewNullablePtr(discord.PermissionBanMembers),
 	Options: []discord.ApplicationCommandOption{
-		discord.ApplicationCommandOptionSubCommand{
-			Name:        "with-message",
-			Description: "Ban a user, sending a message immediately before the ban",
-			Options: []discord.ApplicationCommandOption{
-				discord.ApplicationCommandOptionUser{
-					Name:        "user",
-					Description: "The user to ban",
-					Required:    true,
-				},
-				discord.ApplicationCommandOptionString{
-					Name:        "message",
-					Description: "The message to give the user before banning them (also used as ban reason)",
-					Required:    true,
-				},
-			},
-		},
-
-		discord.ApplicationCommandOptionSubCommand{
-			Name:        "until",
-			Description: "Ban a user from the server for a specified amount of time",
-			Options: []discord.ApplicationCommandOption{
-
-				discord.ApplicationCommandOptionUser{
-					Name:        "user",
-					Description: "The user to ban",
-					Required:    true,
-				},
-				discord.ApplicationCommandOptionString{
-					Name:        "duration",
-					Description: "The duration to ban the user for",
-					Required:    true,
-					Choices:     durationChoices,
-				},
-				discord.ApplicationCommandOptionString{
-					Name:        "reason",
-					Description: "Reason for banning the user",
-					Required:    false,
-				},
-				discord.ApplicationCommandOptionBool{
-					Name:        "send-reason",
-					Description: "Attempt to send the reason to the user as a DM",
-					Required:    false,
-				},
-			},
-		},
+		banWithMessageSubCommand,
+		banUntilSubcommand,
 	},
-}
-
-var durationChoices = []discord.ApplicationCommandOptionChoiceString{
-	{
-		Name:  "1 week",
-		Value: "1w",
-	},
-	{
-		Name:  "2 weeks",
-		Value: "2w",
-	},
-	{
-		Name:  "1 month",
-		Value: "1mo",
-	},
-	{
-		Name:  "3 months",
-		Value: "3mo",
-	},
-	{
-		Name:  "6 months",
-		Value: "6mo",
-	},
-	{
-		Name:  "9 months",
-		Value: "9mo",
-	},
-	{
-		Name:  "1 year",
-		Value: "1y",
-	},
-	{
-		Name:  "2 years",
-		Value: "2y",
-	},
-	{
-		Name:  "3 years",
-		Value: "3y",
-	},
-}
-
-func BanWithMessageHandler(e *handler.CommandEvent) error {
-	utils.LogInteraction("ban with-message", e)
-
-	data := e.SlashCommandInteractionData()
-	user := data.User("user")
-	message := data.String("message")
-
-	return banHandlerInner(e, user, true, message, "")
-}
-
-func BanUntilHandler(e *handler.CommandEvent) error {
-	utils.LogInteraction("ban until", e)
-
-	data := e.SlashCommandInteractionData()
-	user := data.User("user")
-	duration := data.String("duration")
-	reason := data.String("reason")
-	sendReason := data.Bool("send-reason")
-
-	err := banHandlerInner(e, user, sendReason, reason, duration)
-	if err != nil {
-		return err
-	}
-
-	dur, err := utils.ParseLongDuration(duration)
-	if err != nil {
-		return err
-	}
-
-	_, err = model.CreateTempBan(*e.GuildID(), user.ID, e.User().ID, reason, time.Now().Add(dur))
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
 
 func banHandlerInner(e *handler.CommandEvent, user discord.User, sendReason bool, reason, duration string) error {
