+++ interactions/ban/ban_with_message.go
@@ -1,0 +1,35 @@
+package ban
+
+import (
+	"github.com/disgoorg/disgo/discord"
+	"github.com/disgoorg/disgo/handler"
+
+	"github.com/NLLCommunity/heimdallr/utils"
+)
+
+var banWithMessageSubCommand = discord.ApplicationCommandOptionSubCommand{
+	Name:        "with-message",
+	Description: "Ban a user, sending a message immediately before the ban",
+	Options: []discord.ApplicationCommandOption{
+		discord.ApplicationCommandOptionUser{
+			Name:        "user",
+			Description: "The user to ban",
+			Required:    true,
+		},
+		discord.ApplicationCommandOptionString{
+			Name:        "message",
+			Description: "The message to give the user before banning them (also used as ban reason)",
+			Required:    true,
+		},
+	},
+}
+
+func BanWithMessageHandler(e *handler.CommandEvent) error {
+	utils.LogInteraction("ban with-message", e)
+
+	data := e.SlashCommandInteractionData()
+	user := data.User("user")
+	message := data.String("message")
+
+	return banHandlerInner(e, user, true, message, "")
+}
